🎨 Modify : 댓글 삭제 버튼 클릭 후 confirm 알림창 코드 추가

+++ src/components/CommentList/index.jsx
@@ -17,8 +17,10 @@ function CommentList() {
   }, []);
 
   const deleteData = (id) => {
-    dispatch(deleteCommentThunk(id));
-    dispatch(paginationAction(1));
+    if (confirm('삭제하시겠습니까?') === true) {
+      dispatch(deleteCommentThunk(id));
+      dispatch(paginationAction(1));
+    }
   };
   const getData = (id) => {
     if (Object.keys(singleComment).length !== 0) alert('기존 폼을 *수정 완료* 혹은 *수정 취소* 후, 수정 할 댓글을 다시 선택해주세요');
